pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "data_importing_tool"
version = "0.0.1"
authors = [
  { name="Eloi Gabaldon", email="egabaldon@cimne.upc.edu" },
]
description = "A data importing tool"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "beautifulsoup4==4.10.0",
    "bs4==0.0.1",
    "happybase==1.2.0",
    "neo4j==4.4.1",
    "pandas==1.5.1",
    "pymongo==3.11.0",
    "python-dotenv==0.19.2",
    "rdflib==6.2.0",
    "requests~=2.28.1",
    "thefuzz==0.19.0",
    "kafka-python==2.0.2",
    "pycryptodome==3.14.1",
    "python-Levenshtein==0.12.2",
    "html5lib==1.1",
    "openpyxl==3.0.9",
    "xlrd==2.0.1",
    "python-slugify==5.0.2",
    "geopy==2.2.0",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

